Move version throttling tests to own file
